+++ axogen/targets/index.ts
@@ -1,3 +1,4 @@
 export * from "./python.ts";
 export * from "./node.ts";
+export * from "./wasm.ts";
 export * from "./docs.ts";